+++ js/mini-photo-click-handler.js
