+++ js/rendering.js
@@ -1,4 +1,4 @@
-import {onCardClick} from './full-photo-click-handler.js';
+import {onCardClick} from './mini-photo-click-handler.js';
 
 const renderPhotoArray = (photoArray) => {
   const pictureTemplate = document.querySelector('#picture')
